requirements.txt
tkinterdnd2
Pillow
numpy
matplotlib